Skiller ut integrasjonstest av ytelsene til egne tester.
