+++ src/test/kotlin/no/nav/k9brukerdialogapi/ApplicationTest.kt
@@ -1,39 +1,28 @@
 package no.nav.k9brukerdialogapi
 
 import com.github.fppt.jedismock.RedisServer
-import com.github.tomakehurst.wiremock.http.Cookie
 import com.typesafe.config.ConfigFactory
 import io.ktor.config.*
 import io.ktor.http.*
 import io.ktor.server.testing.*
+import io.prometheus.client.CollectorRegistry
 import no.nav.helse.TestUtils.Companion.getAuthCookie
 import no.nav.helse.TestUtils.Companion.getTokenDingsToken
+import no.nav.helse.TestUtils.Companion.requestAndAssert
 import no.nav.helse.dusseldorf.ktor.core.fromResources
 import no.nav.helse.dusseldorf.testsupport.wiremock.WireMockBuilder
-import no.nav.k9brukerdialogapi.SøknadUtils.Companion.søker
 import no.nav.k9brukerdialogapi.wiremock.*
-import no.nav.k9brukerdialogapi.ytelse.Ytelse.*
-import no.nav.k9brukerdialogapi.ytelse.ettersending.domene.Søknadstype
-import no.nav.k9brukerdialogapi.ytelse.fellesdomene.Barn
-import no.nav.k9brukerdialogapi.ytelse.omsorgspengermidlertidigalene.domene.AnnenForelder
-import no.nav.k9brukerdialogapi.ytelse.omsorgspengermidlertidigalene.domene.Situasjon
-import no.nav.k9brukerdialogapi.ytelse.omsorgspengerutvidetrett.domene.SøkerBarnRelasjon
-import no.nav.k9brukerdialogapi.ytelse.omsorgspengerutvidetrett.domene.Søknad
 import org.json.JSONObject
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Nested
-import org.skyscreamer.jsonassert.JSONAssert
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import java.net.URL
-import java.time.LocalDate
 import java.util.*
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
-import no.nav.k9brukerdialogapi.ytelse.omsorgspengermidlertidigalene.domene.Søknad as OmsorgspengerMidlertidigAleneSøknad
 
 class ApplicationTest {
 
@@ -85,6 +74,7 @@ class ApplicationTest {
         @BeforeAll
         @JvmStatic
         fun buildUp() {
+            CollectorRegistry.defaultRegistry.clear()
             engine.start(wait = true)
         }
 
@@ -94,6 +84,7 @@ class ApplicationTest {
             logger.info("Tearing down")
             wireMockServer.stop()
             redisServer.stop()
+            kafkaEnvironment.tearDown()
             logger.info("Tear down complete")
         }
     }
@@ -132,6 +123,8 @@ class ApplicationTest {
         @Test
         fun `Hente søker med loginservice token som cookie`() {
             requestAndAssert(
+                engine = engine,
+                logger = logger,
                 httpMethod = HttpMethod.Get,
                 path = OPPSLAG_URL+SØKER_URL,
                 expectedCode = HttpStatusCode.OK,
@@ -152,6 +145,8 @@ class ApplicationTest {
         @Test
         fun `Hente søker med tokenX token som authorization header`() {
             requestAndAssert(
+                engine = engine,
+                logger = logger,
                 httpMethod = HttpMethod.Get,
                 path = OPPSLAG_URL + SØKER_URL,
                 expectedCode = HttpStatusCode.OK,
@@ -187,6 +182,8 @@ class ApplicationTest {
             )
 
             requestAndAssert(
+                engine = engine,
+                logger = logger,
                 httpMethod = HttpMethod.Get,
                 path = OPPSLAG_URL + SØKER_URL,
                 expectedCode = HttpStatusCode.fromValue(451),
@@ -210,6 +207,8 @@ class ApplicationTest {
         @Test
         fun `Hente søker med tilgangsnivå 3`() {
             requestAndAssert(
+                engine = engine,
+                logger = logger,
                 httpMethod = HttpMethod.Get,
                 path = OPPSLAG_URL + SØKER_URL,
                 cookie = getAuthCookie(fnr = gyldigFødselsnummerA, level = 3),
@@ -226,6 +225,8 @@ class ApplicationTest {
         fun `Test flyt av mellomlagring`() {
             // Legge til mellomlagring
             requestAndAssert(
+                engine = engine,
+                logger = logger,
                 httpMethod = HttpMethod.Post,
                 path = "mellomlagring/OMSORGSPENGER_UTVIDET_RETT",
                 expectedCode = HttpStatusCode.NoContent,
@@ -240,6 +241,8 @@ class ApplicationTest {
 
             // Hente mellomlagring
             requestAndAssert(
+                engine = engine,
+                logger = logger,
                 httpMethod = HttpMethod.Get,
                 path = "mellomlagring/OMSORGSPENGER_UTVIDET_RETT",
                 expectedCode = HttpStatusCode.OK,
@@ -253,6 +256,8 @@ class ApplicationTest {
 
             // Oppdatere mellomlagring
             requestAndAssert(
+                engine = engine,
+                logger = logger,
                 httpMethod = HttpMethod.Put,
                 path = "mellomlagring/OMSORGSPENGER_UTVIDET_RETT",
                 expectedCode = HttpStatusCode.NoContent,
@@ -267,6 +272,8 @@ class ApplicationTest {
 
             // Hente oppdatert mellomlagring
             requestAndAssert(
+                engine = engine,
+                logger = logger,
                 httpMethod = HttpMethod.Get,
                 path = "mellomlagring/OMSORGSPENGER_UTVIDET_RETT",
                 expectedCode = HttpStatusCode.OK,
@@ -280,6 +287,8 @@ class ApplicationTest {
 
             // Slette mellomlagring
             requestAndAssert(
+                engine = engine,
+                logger = logger,
                 httpMethod = HttpMethod.Delete,
                 path = "mellomlagring/OMSORGSPENGER_UTVIDET_RETT",
                 expectedCode = HttpStatusCode.Accepted,
@@ -289,6 +298,8 @@ class ApplicationTest {
 
             // Hente mellomlagring som skal være slettet
             requestAndAssert(
+                engine = engine,
+                logger = logger,
                 httpMethod = HttpMethod.Get,
                 path = "mellomlagring/OMSORGSPENGER_UTVIDET_RETT",
                 expectedCode = HttpStatusCode.OK,
@@ -303,6 +314,8 @@ class ApplicationTest {
         @Test
         fun `Hente barn og eksplisit sjekke at identitetsnummer ikke blir med ved get kall`() {
             val respons = requestAndAssert(
+                engine = engine,
+                logger = logger,
                 httpMethod = HttpMethod.Get,
                 path = OPPSLAG_URL+BARN_URL,
                 expectedCode = HttpStatusCode.OK,
@@ -340,6 +353,8 @@ class ApplicationTest {
         fun `Feil ved henting av barn skal returnere tom liste`() {
             wireMockServer.stubK9OppslagBarn(simulerFeil = true)
             requestAndAssert(
+                engine = engine,
+                logger = logger,
                 httpMethod = HttpMethod.Get,
                 path = OPPSLAG_URL+BARN_URL,
                 expectedCode = HttpStatusCode.OK,
@@ -359,6 +374,8 @@ class ApplicationTest {
         @Test
         fun `Oppslag av alle arbeidsgiver inkludert private og frilansoppdrag`(){
             requestAndAssert(
+                engine = engine,
+                logger = logger,
                 httpMethod = HttpMethod.Get,
                 path = "$OPPSLAG_URL$ARBEIDSGIVER_URL?fra_og_med=2019-01-01&til_og_med=2019-01-30&frilansoppdrag=true&private_arbeidsgivere=true",
                 expectedCode = HttpStatusCode.OK,
@@ -413,6 +430,8 @@ class ApplicationTest {
         @Test
         fun `Oppslag av arbeidsgivere uten private og frilansoppdrag`(){
             requestAndAssert(
+                engine = engine,
+                logger = logger,
                 httpMethod = HttpMethod.Get,
                 path = "$OPPSLAG_URL$ARBEIDSGIVER_URL?fra_og_med=2019-01-01&til_og_med=2019-01-30",
                 expectedCode = HttpStatusCode.OK,
@@ -498,308 +517,5 @@ class ApplicationTest {
                 expectedCode = HttpStatusCode.PayloadTooLarge
             )
         }
-    }
-
-    @Nested
-    inner class OmsorgspengerUtvidetRettTest{
-        @Test
-        fun `Innsending av gyldig søknad`() {
-            val søknad = Søknad(
-                språk = "nb",
-                kroniskEllerFunksjonshemming = true,
-                barn = Barn(
-                    norskIdentifikator = null,
-                    fødselsdato = null,
-                    aktørId = "1000000000001",
-                    navn = "Barn Barnesen"
-                ),
-                relasjonTilBarnet = SøkerBarnRelasjon.FAR,
-                sammeAdresse = true,
-                legeerklæring = listOf(),
-                samværsavtale = listOf(),
-                harBekreftetOpplysninger = true,
-                harForståttRettigheterOgPlikter = true
-            )
-            requestAndAssert(
-                httpMethod = HttpMethod.Post,
-                path = OMSORGSPENGER_UTVIDET_RETT_URL + INNSENDING_URL,
-                expectedCode = HttpStatusCode.Accepted,
-                jwtToken = tokenXToken,
-                requestEntity = søknad.somJson()
-            )
-            val hentet = kafkaKonsumer.hentSøknad(søknad.søknadId, OMSORGSPENGER_UTVIDET_RETT)
-            assertEquals(
-                søknad.tilKomplettSøknad(søker, søknad.tilK9Format(søker)),
-                hentet.data.somOmsorgspengerUtvidetRettKomplettSøknad()
-            )
-        }
-
-        @Test
-        fun `Innsending av ugyldig søknad som får valideringsfeil`() {
-            val søknad = Søknad(
-                språk = "nb",
-                kroniskEllerFunksjonshemming = true,
-                barn = Barn(
-                    norskIdentifikator = "123",
-                    navn = "Barn Barnesen"
-                ),
-                relasjonTilBarnet = SøkerBarnRelasjon.FAR,
-                sammeAdresse = true,
-                legeerklæring = listOf(),
-                samværsavtale = listOf(),
-                harBekreftetOpplysninger = false,
-                harForståttRettigheterOgPlikter = false
-            ).somJson()
-
-            requestAndAssert(
-                httpMethod = HttpMethod.Post,
-                path = OMSORGSPENGER_UTVIDET_RETT_URL + INNSENDING_URL,
-                expectedCode = HttpStatusCode.BadRequest,
-                jwtToken = tokenXToken,
-                expectedResponse = """
-                {
-                  "detail": "Requesten inneholder ugyldige paramtere.",
-                  "instance": "about:blank",
-                  "type": "/problem-details/invalid-request-parameters",
-                  "title": "invalid-request-parameters",
-                  "invalid_parameters": [
-                    {
-                      "name": "harForståttRettigheterOgPlikter",
-                      "reason": "Må ha forstått rettigheter og plikter for å sende inn søknad.",
-                      "invalid_value": null,
-                      "type": "entity"
-                    }, {
-                      "name": "harBekreftetOpplysninger",
-                      "reason": "Opplysningene må bekreftes for å sende inn søknad.",
-                      "invalid_value": null,
-                      "type": "entity"
-                    }, {
-                      "type": "entity",
-                      "name": "barn.norskIdentifikator",
-                      "reason": "Ikke gyldig norskIdentifikator.",
-                      "invalid_value": "123"
-                    }
-                  ],
-                  "status": 400
-                }
-            """.trimIndent(),
-                requestEntity = søknad
-            )
-        }
-    }
-
-    @Nested
-    inner class OmsorgspengerMidlertidigAleneTest{
-        @Test
-        fun `Innsending av gyldig søknad`() {
-            val søknad = OmsorgspengerMidlertidigAleneSøknad(
-                id = "123456789",
-                språk = "nb",
-                annenForelder = AnnenForelder(
-                    navn = "Berit",
-                    fnr = "02119970078",
-                    situasjon = Situasjon.FENGSEL,
-                    situasjonBeskrivelse = "Sitter i fengsel..",
-                    periodeOver6Måneder = false,
-                    periodeFraOgMed = LocalDate.parse("2020-01-01"),
-                    periodeTilOgMed = LocalDate.parse("2020-10-01")
-                ),
-                barn = listOf(
-                    Barn(
-                        navn = "Ole Dole",
-                        norskIdentifikator = "25058118020",
-                        aktørId = null
-                    )
-                ),
-                harBekreftetOpplysninger = true,
-                harForståttRettigheterOgPlikter = true
-            )
-            requestAndAssert(
-                httpMethod = HttpMethod.Post,
-                path = OMSORGSPENGER_MIDLERTIDIG_ALENE_URL + INNSENDING_URL,
-                expectedCode = HttpStatusCode.Accepted,
-                jwtToken = tokenXToken,
-                expectedResponse = null,
-                requestEntity = søknad.somJson()
-            )
-            val hentet = kafkaKonsumer.hentSøknad(søknad.søknadId, OMSORGSPENGER_MIDLERTIDIG_ALENE)
-            assertEquals(
-                søknad.tilKomplettSøknad(søker, søknad.tilK9Format(søker)),
-                hentet.data.somOmsorgspengerMidlertidigAleneKomplettSøknad()
-            )
-        }
-
-        @Test
-        fun `Innsending av ugyldig søknad som får valideringsfeil`() {
-            val søknad = OmsorgspengerMidlertidigAleneSøknad(
-                id = "123456789",
-                språk = "nb",
-                annenForelder = AnnenForelder(
-                    navn = "Berit",
-                    fnr = "ikke gyldig",
-                    situasjon = Situasjon.FENGSEL,
-                    situasjonBeskrivelse = "Sitter i fengsel..",
-                    periodeOver6Måneder = false,
-                    periodeFraOgMed = LocalDate.parse("2020-01-01"),
-                    periodeTilOgMed = null
-                ),
-                barn = listOf(
-                    Barn(
-                        navn = "Ole Dole",
-                        norskIdentifikator = "ikke gyldig",
-                        aktørId = null
-                    )
-                ),
-                harBekreftetOpplysninger = false,
-                harForståttRettigheterOgPlikter = true
-            ).somJson()
-
-            requestAndAssert(
-                httpMethod = HttpMethod.Post,
-                path = OMSORGSPENGER_MIDLERTIDIG_ALENE_URL + INNSENDING_URL,
-                expectedCode = HttpStatusCode.BadRequest,
-                jwtToken = tokenXToken,
-                expectedResponse = """
-                    {
-                      "type": "/problem-details/invalid-request-parameters",
-                      "title": "invalid-request-parameters",
-                      "status": 400,
-                      "detail": "Requesten inneholder ugyldige paramtere.",
-                      "instance": "about:blank",
-                      "invalid_parameters": [
-                        {
-                          "type": "entity",
-                          "name": "harBekreftetOpplysninger",
-                          "reason": "Opplysningene må bekreftes for å sende inn søknad.",
-                          "invalid_value": null
-                        }, {
-                          "type": "entity",
-                          "name": "AnnenForelder.fnr",
-                          "reason": "Fødselsnummer på annen forelder må være gyldig norsk identifikator",
-                          "invalid_value": "ikke gyldig"
-                        },
-                        {
-                          "type": "entity",
-                          "name": "AnnenForelder.periodeTilOgMed",
-                          "reason": "periodeTilOgMed kan ikke være null dersom situasjonen er FENGSEL",
-                          "invalid_value": null
-                        },
-                        {
-                          "type": "entity",
-                          "name": "barn.norskIdentifikator",
-                          "reason": "Ikke gyldig norskIdentifikator.",
-                          "invalid_value": "ikke gyldig"
-                        }
-                      ]
-                    }
-            """.trimIndent(),
-                requestEntity = søknad
-            )
-        }
-    }
-
-    @Nested
-    inner class EttersendingTest{
-        @Test
-        fun `Innsending av gyldig søknad`() {
-            val vedlegg = URL(engine.jpegUrl(jwtToken = tokenXToken))
-            val søknad = no.nav.k9brukerdialogapi.ytelse.ettersending.domene.Søknad(
-                språk = "nb",
-                vedlegg = setOf<URL>(vedlegg).toList(),
-                beskrivelse = "Sykt barn...",
-                søknadstype = Søknadstype.PLEIEPENGER_SYKT_BARN,
-                harBekreftetOpplysninger = true,
-                harForståttRettigheterOgPlikter = true
-            )
-            requestAndAssert(
-                httpMethod = HttpMethod.Post,
-                path = ETTERSENDING_URL + INNSENDING_URL,
-                expectedCode = HttpStatusCode.Accepted,
-                jwtToken = tokenXToken,
-                expectedResponse = null,
-                requestEntity = søknad.somJson()
-            )
-            val hentet = kafkaKonsumer.hentSøknad(søknad.søknadId, ETTERSENDING)
-            assertEquals(
-                søknad.somKomplettSøknad(søker, søknad.somK9Format(søker), listOf("nav-logo.png")),
-                hentet.data.somEttersendingKomplettSøknad()
-            )
-        }
-
-        @Test
-        fun `Innsending av ugyldig søknad som får valideringsfeil`() {
-            val søknad = no.nav.k9brukerdialogapi.ytelse.ettersending.domene.Søknad(
-                språk = "nb",
-                vedlegg = listOf(),
-                søknadstype = Søknadstype.PLEIEPENGER_SYKT_BARN,
-                beskrivelse = null,
-                harBekreftetOpplysninger = true,
-                harForståttRettigheterOgPlikter = true
-            )
-
-            requestAndAssert(
-                httpMethod = HttpMethod.Post,
-                path = ETTERSENDING_URL + INNSENDING_URL,
-                expectedCode = HttpStatusCode.BadRequest,
-                jwtToken = tokenXToken,
-                expectedResponse = """
-                    {
-                      "detail": "Requesten inneholder ugyldige paramtere.",
-                      "instance": "about:blank",
-                      "type": "/problem-details/invalid-request-parameters",
-                      "title": "invalid-request-parameters",
-                      "invalid_parameters": [
-                        {
-                          "name": "vedlegg",
-                          "reason": "Liste over vedlegg kan ikke være tom.",
-                          "invalid_value": [],
-                          "type": "entity"
-                        },
-                        {
-                          "type": "entity",
-                          "name": "beskrivelse",
-                          "invalid_value" : null,
-                          "reason": "Beskrivelse kan ikke være tom, null eller blank dersom det gjelder pleiepenger."
-                        }
-                      ],
-                      "status": 400
-                    }
-            """.trimIndent(),
-                requestEntity = søknad.somJson()
-            )
-        }
-    }
-
-    fun requestAndAssert(
-        httpMethod: HttpMethod,
-        path: String,
-        requestEntity: String? = null,
-        expectedResponse: String? = null,
-        expectedCode: HttpStatusCode,
-        jwtToken: String? = null,
-        cookie: Cookie? = null
-    ) : String? {
-        val respons: String?
-        with(engine) {
-            handleRequest(httpMethod, path) {
-                if (cookie != null) addHeader(HttpHeaders.Cookie, cookie.toString())
-                if (jwtToken != null) addHeader(HttpHeaders.Authorization, "Bearer $jwtToken")
-                logger.info("Request Entity = $requestEntity")
-                addHeader(HttpHeaders.Accept, "application/json")
-                if (requestEntity != null) addHeader(HttpHeaders.ContentType, "application/json")
-                if (requestEntity != null) setBody(requestEntity)
-            }.apply {
-                logger.info("Response Entity = ${response.content}")
-                logger.info("Expected Entity = $expectedResponse")
-                respons = response.content
-                assertEquals(expectedCode, response.status())
-                if (expectedResponse != null) {
-                    JSONAssert.assertEquals(expectedResponse, response.content!!, true)
-                } else {
-                    assertEquals(expectedResponse, response.content)
-                }
-            }
-        }
-        return respons
     }
 }
