+++ src/test/kotlin/no/nav/k9brukerdialogapi/ytelse/omsorgspengerutvidetrett/OmsorgspengerUtvidetRettTest.kt
@@ -1,0 +1,171 @@
+package no.nav.k9brukerdialogapi.ytelse.omsorgspengerutvidetrett
+
+import com.github.fppt.jedismock.RedisServer
+import com.typesafe.config.ConfigFactory
+import io.ktor.config.*
+import io.ktor.http.*
+import io.ktor.server.testing.*
+import io.prometheus.client.CollectorRegistry
+import no.nav.helse.TestUtils
+import no.nav.helse.TestUtils.Companion.requestAndAssert
+import no.nav.helse.dusseldorf.testsupport.wiremock.WireMockBuilder
+import no.nav.k9brukerdialogapi.*
+import no.nav.k9brukerdialogapi.wiremock.k9BrukerdialogApiConfig
+import no.nav.k9brukerdialogapi.wiremock.stubK9OppslagBarn
+import no.nav.k9brukerdialogapi.wiremock.stubK9OppslagSoker
+import no.nav.k9brukerdialogapi.ytelse.Ytelse
+import no.nav.k9brukerdialogapi.ytelse.fellesdomene.Barn
+import no.nav.k9brukerdialogapi.ytelse.omsorgspengerutvidetrett.domene.SøkerBarnRelasjon
+import no.nav.k9brukerdialogapi.ytelse.omsorgspengerutvidetrett.domene.Søknad
+import org.junit.jupiter.api.AfterAll
+import org.junit.jupiter.api.BeforeAll
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class OmsorgspengerUtvidetRettTest {
+
+    private companion object{
+        private val logger: Logger = LoggerFactory.getLogger(OmsorgspengerUtvidetRettTest::class.java)
+        val wireMockServer = WireMockBuilder()
+            .withAzureSupport()
+            .withNaisStsSupport()
+            .withLoginServiceSupport()
+            .withTokendingsSupport()
+            .k9BrukerdialogApiConfig()
+            .build()
+            .stubK9OppslagSoker()
+            .stubK9OppslagBarn()
+
+        private val kafkaEnvironment = KafkaWrapper.bootstrap()
+        private val kafkaKonsumer = kafkaEnvironment.testConsumer()
+
+        private val gyldigFødselsnummerA = "02119970078"
+        private val tokenXToken = TestUtils.getTokenDingsToken(fnr = gyldigFødselsnummerA)
+
+        val redisServer: RedisServer = RedisServer
+            .newRedisServer().apply { start() }
+
+        fun getConfig(): ApplicationConfig {
+            val fileConfig = ConfigFactory.load()
+            val testConfig = ConfigFactory.parseMap(
+                TestConfiguration.asMap(
+                    wireMockServer = wireMockServer,
+                    kafkaEnvironment = kafkaEnvironment,
+                    redisServer = redisServer
+                )
+            )
+            val mergedConfig = testConfig.withFallback(fileConfig)
+
+            return HoconApplicationConfig(mergedConfig)
+        }
+
+        val engine = TestApplicationEngine(createTestEnvironment { config = getConfig() })
+
+        @BeforeAll
+        @JvmStatic
+        fun buildUp() {
+            CollectorRegistry.defaultRegistry.clear()
+            engine.start(wait = true)
+        }
+
+        @AfterAll
+        @JvmStatic
+        fun tearDown() {
+            wireMockServer.stop()
+            redisServer.stop()
+            kafkaEnvironment.tearDown()
+        }
+    }
+
+    @Test
+    fun `Innsending av gyldig søknad`() {
+        val søknad = Søknad(
+            språk = "nb",
+            kroniskEllerFunksjonshemming = true,
+            barn = Barn(
+                norskIdentifikator = null,
+                fødselsdato = null,
+                aktørId = "1000000000001",
+                navn = "Barn Barnesen"
+            ),
+            relasjonTilBarnet = SøkerBarnRelasjon.FAR,
+            sammeAdresse = true,
+            legeerklæring = listOf(),
+            samværsavtale = listOf(),
+            harBekreftetOpplysninger = true,
+            harForståttRettigheterOgPlikter = true
+        )
+        requestAndAssert(
+            engine = engine,
+            logger = logger,
+            httpMethod = HttpMethod.Post,
+            path = OMSORGSPENGER_UTVIDET_RETT_URL + INNSENDING_URL,
+            expectedCode = HttpStatusCode.Accepted,
+            jwtToken = tokenXToken,
+            requestEntity = søknad.somJson()
+        )
+        val hentet = kafkaKonsumer.hentSøknad(søknad.søknadId, Ytelse.OMSORGSPENGER_UTVIDET_RETT)
+        assertEquals(
+            søknad.tilKomplettSøknad(SøknadUtils.søker, søknad.tilK9Format(SøknadUtils.søker)),
+            hentet.data.somOmsorgspengerUtvidetRettKomplettSøknad()
+        )
+    }
+
+    @Test
+    fun `Innsending av ugyldig søknad som får valideringsfeil`() {
+        val søknad = Søknad(
+            språk = "nb",
+            kroniskEllerFunksjonshemming = true,
+            barn = Barn(
+                norskIdentifikator = "123",
+                navn = "Barn Barnesen"
+            ),
+            relasjonTilBarnet = SøkerBarnRelasjon.FAR,
+            sammeAdresse = true,
+            legeerklæring = listOf(),
+            samværsavtale = listOf(),
+            harBekreftetOpplysninger = false,
+            harForståttRettigheterOgPlikter = false
+        ).somJson()
+
+        requestAndAssert(
+            engine = engine,
+            logger = logger,
+            httpMethod = HttpMethod.Post,
+            path = OMSORGSPENGER_UTVIDET_RETT_URL + INNSENDING_URL,
+            expectedCode = HttpStatusCode.BadRequest,
+            jwtToken = tokenXToken,
+            expectedResponse = """
+                {
+                  "detail": "Requesten inneholder ugyldige paramtere.",
+                  "instance": "about:blank",
+                  "type": "/problem-details/invalid-request-parameters",
+                  "title": "invalid-request-parameters",
+                  "invalid_parameters": [
+                    {
+                      "name": "harForståttRettigheterOgPlikter",
+                      "reason": "Må ha forstått rettigheter og plikter for å sende inn søknad.",
+                      "invalid_value": null,
+                      "type": "entity"
+                    }, {
+                      "name": "harBekreftetOpplysninger",
+                      "reason": "Opplysningene må bekreftes for å sende inn søknad.",
+                      "invalid_value": null,
+                      "type": "entity"
+                    }, {
+                      "type": "entity",
+                      "name": "barn.norskIdentifikator",
+                      "reason": "Ikke gyldig norskIdentifikator.",
+                      "invalid_value": "123"
+                    }
+                  ],
+                  "status": 400
+                }
+            """.trimIndent(),
+            requestEntity = søknad
+        )
+    }
+
+}
